Modify code to be the relationship Many-to-Many relationship

+++ src/main/java/com/ProjectManagementBoardAPI/ProjectManagementBoardAPI/Repositories/CardRepository.java
@@ -3,5 +3,5 @@ package com.ProjectManagementBoardAPI.ProjectManagementBoardAPI.Repositories;
 import com.ProjectManagementBoardAPI.ProjectManagementBoardAPI.Models.Card;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CardRepository extends JpaRepository<Card, Integer> {
+public interface CardRepository extends JpaRepository<Card, Long> {
 }
